Split out addTermForPost
